requirements.txt
Django==3.2.5
django-extensions==3.1.3
django-group-by==0.3.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary>=2.8,<2.9
